+++ dynaml/addition_test.go
@@ -12,7 +12,7 @@ var _ = Describe("addition", func() {
 			IntegerExpr{3},
 		}
 
-		Expect(expr).To(EvaluateAs(5, FakeBinding{}))
+		Expect(expr).To(EvaluateAs(int64(5), FakeBinding{}))
 	})
 
 	Context("when the left-hand side is not an integer", func() {
